requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.4
Django==3.1.6
django-countries==7.0
django-crispy-forms==1.11.1
django-debug-toolbar==3.2
django-js-asset==1.2.2
django-mptt==0.12.0
django-phonenumber-field==5.0.0
flake8==3.8.4
flake8-isort==4.0.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
MouseInfo==0.1.3
mypy-extensions==0.4.3
pathspec==0.8.1
phonenumbers==8.12.19
Pillow==8.1.0
PyAutoGUI==0.9.52
pycodestyle==2.6.0
pyflakes==2.2.0
PyGetWindow==0.0.9
pylint==2.6.2
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
pytz==2021.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
stripe==2.56.0
testfixtures==6.17.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wrapt==1.12.1
